requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
langchain==0.0.350
chromadb==0.4.18
pydantic==2.5.2
python-multipart==0.0.6 